requirements.txt
pandas
numpy
requests
yfinance
python-dotenv
snowflake-connector-python
openai
beautifulsoup4
lxml
